+++ library/src/main/java/com/klinker/android/simple_videoview/SimpleVideoView.java
@@ -16,34 +16,32 @@ package com.klinker.android.simple_videoview;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.graphics.Point;
 import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.net.Uri;
 import android.util.AttributeSet;
-import android.view.Display;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.WindowManager;
 import android.widget.LinearLayout;
-
-import java.io.IOException;
+import android.widget.RelativeLayout;
 
 /**
  * VideoView implementation that simplifies things, fixes aspect ratio, and allows 
  * you to specify whether or not you want to overtake the system audio.
  */
-public class SimpleVideoView extends LinearLayout {
+public class SimpleVideoView extends RelativeLayout {
 
     private MediaPlayer mediaPlayer;
 
     private SurfaceView surfaceView;
     private SurfaceHolder surfaceHolder;
-    private View progressBar;
+    private LinearLayout progressBar;
+
+    private VideoPlaybackErrorTracker errorTracker;
 
     private boolean loop = false;
     private boolean stopSystemAudio = false;
@@ -82,7 +80,7 @@ public class SimpleVideoView extends LinearLayout {
      */
     private void init() {
         // add a progress spinner
-        progressBar = LayoutInflater.from(getContext()).inflate(R.layout.progress_bar, null, false);
+        progressBar = (LinearLayout) LayoutInflater.from(getContext()).inflate(R.layout.progress_bar, this, false);
         addView(progressBar);
 
         setGravity(Gravity.CENTER);
@@ -100,9 +98,7 @@ public class SimpleVideoView extends LinearLayout {
         mediaPlayer = new MediaPlayer();
         mediaPlayer.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
             @Override
-            public void onPrepared(MediaPlayer mediaPlayer) {
-                progressBar.setVisibility(View.GONE);
-
+            public void onPrepared(final MediaPlayer mediaPlayer) {
                 scalePlayer();
 
                 if (stopSystemAudio) {
@@ -114,9 +110,42 @@ public class SimpleVideoView extends LinearLayout {
                     mediaPlayer.setVolume(0, 0);
                 }
 
+                progressBar.setVisibility(View.GONE);
+
                 mediaPlayer.setDisplay(surfaceHolder);
                 mediaPlayer.setLooping(loop);
                 mediaPlayer.start();
+            }
+        });
+
+        mediaPlayer.setOnErrorListener(new MediaPlayer.OnErrorListener() {
+            /**
+             * Called to indicate an error.
+             *
+             * @param mp      the MediaPlayer the error pertains to
+             * @param what    the type of error that has occurred:
+             * <ul>
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_UNKNOWN}
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_SERVER_DIED}
+             * </ul>
+             * @param extra an extra code, specific to the error. Typically
+             * implementation dependent.
+             * <ul>
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_IO}
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_MALFORMED}
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_UNSUPPORTED}
+             * <li>{@link android.media.MediaPlayer.OnErrorListener#MEDIA_ERROR_TIMED_OUT}
+             * <li><code>MEDIA_ERROR_SYSTEM (-2147483648)</code> - low-level system error.
+             * </ul>
+             */
+            @Override
+            public boolean onError(MediaPlayer mp, int what, int extra) {
+                if (errorTracker != null) {
+                    errorTracker.onPlaybackError(
+                            new RuntimeException("Error playing video! what code: " + what + ", extra code: " + extra)
+                    );
+                }
+                return true;
             }
         });
         mediaPlayer.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
@@ -131,12 +160,10 @@ public class SimpleVideoView extends LinearLayout {
             }
         });
 
-        LinearLayout.LayoutParams surfaceViewParams =
+        final RelativeLayout.LayoutParams surfaceViewParams =
                 new LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT);
         surfaceView = new SurfaceView(getContext());
         surfaceView.setLayoutParams(surfaceViewParams);
-
-        addView(surfaceView);
 
         surfaceHolder = surfaceView.getHolder();
         surfaceHolder.addCallback(new SurfaceHolder.Callback() {
@@ -144,14 +171,27 @@ public class SimpleVideoView extends LinearLayout {
             @Override public void surfaceDestroyed(SurfaceHolder surfaceHolder) { }
             @Override
             public void surfaceCreated(SurfaceHolder surfaceHolder) {
-                try {
-                    mediaPlayer.setDataSource(getContext(), videoUri);
-                    mediaPlayer.prepare();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
+                new Thread(new Runnable() {
+                    @Override
+                    public void run() {
+                        try {
+                            // this needs to be run on a background thread.
+                            // set data source can take upwards of 1-2 seconds
+                            mediaPlayer.setDataSource(getContext(), videoUri);
+                            mediaPlayer.prepare();
+                        } catch (Exception e) {
+                            if (errorTracker != null) {
+                                errorTracker.onPlaybackError(e);
+                            }
+                        }
+                    }
+                }).start();
+
             }
         });
+
+        addView(surfaceView, 0);
+
     }
 
     /**
@@ -172,7 +212,6 @@ public class SimpleVideoView extends LinearLayout {
             lp.width = (int) (videoProportion * (float) getHeight());
             lp.height = getHeight();
         }
-
 
         surfaceView.setLayoutParams(lp);
     }
@@ -258,5 +297,18 @@ public class SimpleVideoView extends LinearLayout {
         } catch (Exception e) {
             return false;
         }
+    }
+
+    /**
+     * Will return a result if there is an error playing the video
+     *
+     * @param tracker
+     */
+    public void setErrorTracker(VideoPlaybackErrorTracker tracker) {
+        this.errorTracker = tracker;
+    }
+
+    public interface VideoPlaybackErrorTracker {
+        void onPlaybackError(Exception e);
     }
 }