+++ changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 1.1.0:
+- Fix performance issues by setting the data source from a background thread
+- Improve the loading UI
+- Create an error listener interface
+
 ### Version 1.0.4:
 - Catch an exception